Update define_projections - declare CRS:: and document.
import was not declared properly, now I use the :: and documentation
was increased such that others might understand what is going on.
The objects are still assigned to the global environment, and in the
future it might be better to assign them to an environment and then
export just that environment, or provide an additional argument where
people could specify  the environment that they want.